Fixes Contentful sync. (#547)
### Description

Continuation of Edge GA renaming (#532).

This PR gets us back into a good and working state. It accomplishes
three things.

1. Since the Contentful sync is unidirectional (vs. bi-directional),
templates in the repo were out of sync with Contentful. This gets things
back into a good place. General rule of thumb... ***Repo centric
templates should never be changed in Contentful directly or changes will
get overwritten.***
2. Updated log messages to contain the template slug, making it easier
to understand what is happening and when since everything is async. This
really helps with troubleshooting errors as the thrown stacktrace might
not be from the last template displayed.
3. The `update-templates` command, which is used by GitHub Actions, only
updates changed files on a PR. This is typically good, but was unhelpful
in this case as so many templates were out of sync and pushing
subsequent PRs to the repo wouldn't see past commits as changes. To
circumvent this, I created a `force-update-templates` rpm script, which
will force all templates to be synced. This will not be hooked up to
GitHub Actions, but can instead be run locally, if needed.

### Type of Change

- [x] Example updates (Bug fixes, new features, etc.)
- [x] Other (changes to the codebase, but not to examples)