Update dependencies and fix TypeScript errors at Blob Starter (#867)
### Description

This PR includes several updates to improve the project's overall health
and resolve TypeScript errors.

* Updated all dependencies in package.json to their latest versions.
* Replaced the deprecated BlobResult import in vercel/blob with
PutBlobResult.
* Updated the t parameter in the toast function call with the correct
type.

## Why is it important?

The update from `BlobResult` to `PutBlobResult` in vercel/blob is a
crucial change. BlobResult is deprecated, and continuing to use it may
lead to unexpected behavior or failures in the future as further updates
and changes are made to the vercel/blob package.

New users can confidently use the template as a reference or starting
point, knowing that it follows the latest practices and uses the most
recent and supported features. This makes their introduction to
vercel/blob smoother and more enjoyable, and helps them learn and adopt
best practices from the start.

### Type of Change

- [ ] New Example
- [X] Example updates (Bug fixes, new features, etc.)
- [ ] Other (changes to the codebase, but not to examples)